Version 11 - now accepts any form of "male" or "female" to decide name
The system will look up the user's inputted gender in comparison of a list of gender words from multiple languages. When a term is identified (like  the term"femme" is French for "woman") it will gender the name to match it.  This gives greater flexibility of inputs while still aligning the name to the gender.  

Everyone is assumed non-binary (by which all names are valid) unless specifically identified as male or female by which a gendered name is used.